+++ CHANGELOG.md
@@ -21,6 +21,7 @@ BUG FIXES:
   - Show a human-friendly error if VBoxManage is not found (exit
     status 126). [GH-934]
   - Action hook prepend/append will only prepend or append once.
+  - Retry SSH on Errno::EACCES.
 
 ## 1.1.4 (March 25, 2013)